@@ -4,11 +4,12 @@ description: A common library for charting packages.
 homepage: https://github.com/google/charts
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: '>=3.3.0'
 
 dependencies:
   collection: ^1.14.5
-  intl: ">=0.15.2 < 0.18.0"
+  intl: ^0.19.0
   logging: ^1.0.2
   meta: ^1.1.1
   vector_math: ^2.0.8
