@@ -4,8 +4,8 @@ description: Material Design charting library for flutter.
 homepage: https://github.com/google/charts
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: '>=2.5.0'
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: '>=3.3.0'
 
 dependencies:
   # Pointing this to a local path allows for pointing to the latest code
@@ -13,20 +13,19 @@ dependencies:
   #
   # The pub version of charts_flutter will point to the pub version of charts_common.
   # The latest pub version is commented out and shown below as an example.
-  charts_common: ^0.12.0
-  # charts_common:
-  #   path: ../charts_common/
+  # charts_common: ^0.12.0
+  charts_common:
+    path: ../charts_common/
   collection: ^1.14.5
   flutter:
     sdk: flutter
-  intl: ">=0.15.2 < 0.18.0"
+  intl: ^0.19.0
   logging: ^1.0.2
   meta: ^1.1.1
 
-
 dev_dependencies:
   mockito: ^5.0.0
-  build_runner: ^1.11.0
+  build_runner: ^2.0.6
   flutter_test:
     sdk: flutter
   test: ^1.3.0
