@@ -45,7 +45,7 @@ abstract class TimeStepper {
 
 /// Iterator with a reset function that can be used multiple times to avoid
 /// object instantiation during the Android layout/draw phases.
-abstract class TimeStepIterator extends Iterator<DateTime> {
+abstract class TimeStepIterator implements Iterator<DateTime> {
   /// Reset the iterator and set the tickIncrement to the specified value.
   ///
   /// This method is provided so that the same iterator instance can be used for
